WIP: abstracting server code and organizing index.js in regions
